feat(fuse.js): rename folder from fuse to fuse.js
